requirements.txt
asgiref==3.8.1
asttokens==2.4.1
attrs==24.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
begins==0.9
bleach==6.1.0
cachetools==5.4.0
cattrs==23.2.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
cors==1.0.1
cryptography==42.0.8
decorator==5.1.1
defusedxml==0.8.0rc2
dj-rest-auth==6.0.0
Django==5.0.6
django-allauth==0.63.6
django-cors-headers==4.4.0
django-erd==1.0
django-extensions==3.2.3
django-model-utils==4.5.1
django-schema-graph==3.1.0
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
docopt==0.6.2
drf-yasg==1.21.7
executing==2.0.1
fastjsonschema==2.20.0
filelock==3.15.4
future==1.0.0
gevent==24.2.1
google-api-core==2.19.1
google-api-python-client==2.138.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.63.2
graphviz==0.20.3
greenlet==3.0.3
httplib2==0.22.0
idna==3.7
inflection==0.5.1
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mistune==3.0.2
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
oauthlib==3.2.2
packaging==24.1
pandocfilters==1.5.1
parso==0.8.4
pickleshare==0.7.5
pillow==10.3.0
pip-autoremove==0.10.0
pipdeptree==2.23.1
pipreqs==0.5.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
proto-plus==1.24.0
gunicorn==20.1.0
protobuf==5.27.2
# psycopg2==2.9.9
pure_eval==0.2.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydotplus==2.0.2
Pygments==2.18.0
PyJWT==2.8.0
pyOpenSSL==24.2.1
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python3-openid==3.2.0
pytz==2024.1
# pywin32==306
whitenoise==6.0.0
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==2.0.0
rpds-py==0.20.0
rsa==4.9
six==1.16.0
social-auth-app-django==5.4.1
social-auth-core==4.5.4
soupsieve==2.6
sqlparse==0.5.0
stack-data==0.6.3
tinycss2==1.3.0
tldextract==5.1.2
tornado==6.4.1
traitlets==5.14.3
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.3
yarg==0.1.9
zope.event==5.0
zope.interface==6.4.post2
psycopg2-binary==2.9.9
vobject==0.9.6.1
sentry-sdk==2.14.0
